:construction::art::package: Wind graph initial setup
