+++ pdill.dev/package.json
@@ -12,6 +12,7 @@
     "@material-ui/core": "^4.10.1",
     "@material-ui/icons": "^4.9.1",
     "axios": "^0.21.1",
+    "chart.js": "^2.9.4",
     "dotenv-webpack": "^6.0.0",
     "harmony": "^1.5.1",
     "luxon": "^1.25.0",
@@ -20,6 +21,7 @@
     "next-images": "^1.4.0",
     "next-progressbar": "^1.0.0",
     "react": "16.13.1",
+    "react-chartjs-2": "^2.11.1",
     "react-dom": "16.13.1",
     "react-ga": "^3.1.2",
     "react-icons": "^3.10.0"
